Adding broadcast hint as a parameter to FetchManager
This hint must be consistent with the one used by content server, but it
was not the case before (and was defined independently twice anyways).

Another change in this commit: broadcast forwarding hint does not
contain application name anymore. Not sure if we need to return it back
or not.

Change-Id: I685620ba1ec74a641015c91f1d6e7b5ff5969215